+++ forwarder/metrics/__init__.py
@@ -1,0 +1,45 @@
+"""
+Classes to model monitoring metrics based on the Prometheus client library.
+
+Since we push these metrics to Graphite instead of relying on Prometheus
+standard mechanisms, we provide adapter classes to prevent coupling other
+modules with certain Prometheus internals.
+"""
+
+from prometheus_client import Counter as PrometheusCounter
+from prometheus_client import Gauge as PrometheusGauge
+from prometheus_client import Summary as PrometheusSummary
+
+
+class Gauge(PrometheusGauge):
+    @property
+    def name(self):
+        return self._name
+
+    @property
+    def value(self):
+        return self._value.get()
+
+
+class Counter(PrometheusCounter):
+    @property
+    def name(self):
+        return self._name
+
+    @property
+    def value(self):
+        return self._value.get()
+
+
+class Summary(PrometheusSummary):
+    @property
+    def name(self):
+        return self._name
+
+    @property
+    def count(self):
+        return self._count.get()
+
+    @property
+    def sum(self):
+        return self._sum.get()